move Screen post to feature folder
